Create directories with separated components #12
