+++ CMakeLists.txt
@@ -17,29 +17,6 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 find_package(Qt5 COMPONENTS Widgets REQUIRED)
 find_package(GTest MODULE REQUIRED)
 
-add_library(my_bsp
-    player.cpp
-    player.h
-    polygon.cpp
-    polygon.h)
-target_link_libraries(my_bsp PUBLIC Qt5::Widgets)
-
-add_executable(bsp_tool
-  main.cpp
-  mainwindow.cpp
-  mainwindow.h
-  mainwindow.ui
-  ut_bsp_tool.qrc
-)
-target_link_libraries(bsp_tool PRIVATE my_bsp Qt5::Widgets)
-
-add_executable(create_world
-    create_world.cpp
-    player.cpp
-    player.h
-)
-target_link_libraries(create_world PRIVATE my_bsp Qt5::Widgets)
-
-add_executable(tests tests.cpp)
-target_link_libraries(tests PRIVATE GTest::GTest GTest::Main)
-add_test(NAME tests COMMAND tests)
+add_subdirectory(bsp)
+add_subdirectory(apps)
+add_subdirectory(tests)
