+++ bsp/polygon.cpp
